+++ components/split/include/split_sync.h
@@ -56,5 +56,5 @@ bool split_sync_remote_matrix_changed(void);
 esp_err_t split_sync_send_full_state(const uint8_t *peer_mac,
                                       const uint8_t *matrix,
                                       uint8_t active_layer,
-                                      uint16_t seq);
+                                      uint64_t seq);
